+++ src/hyperiontf/api/rest_client/client.py
@@ -55,6 +55,13 @@ class Client:
         :param connection_timeout: The timeout for establishing a connection.
         :param request_timeout: The timeout for completing a request.
         """
+        self.scheme: Optional[str]
+        self.netloc: Optional[str]
+        self.path: Optional[str]
+        self.params: Optional[str]
+        self.query: Optional[str]
+        self.fragment: Optional[str]
+
         self.scheme, self.netloc, self.path, self.params, self.query, self.fragment = (
             self._parse_url(url)
             if url