+++ tests/install_web_dependencies.sh
@@ -12,3 +12,4 @@ pip install opencv-python-headless
 pip install opencv-contrib-python
 pip install ptyprocess
 pip install paramiko
+pip install lxml